NSA-7274 - Moved helper functions to utils folder
