+++ src/utils/index.js
@@ -1,10 +1,14 @@
 const deprecate = require('./deprecateMiddleware');
 const encodeNumberToString = require('./encodeNumberToString');
 const { mapAsync, forEachAsync } = require('./asyncHelpers');
+const { mapArrayToProperty, arrayToMapById, mapAndFilterArray } = require('./helperFunctions');
 
 module.exports = {
   deprecate,
   encodeNumberToString,
   mapAsync,
   forEachAsync,
+  mapArrayToProperty,
+  arrayToMapById,
+  mapAndFilterArray
 };
